+++ packages/react-dom/src/client/ReactDOM.js
@@ -35,6 +35,7 @@ import {
   attemptUserBlockingHydration,
   attemptContinuousHydration,
   attemptHydrationAtCurrentPriority,
+  runWithPriority,
 } from 'react-reconciler/src/ReactFiberReconciler';
 import {createPortal as createPortalImpl} from 'react-reconciler/src/ReactPortal';
 import {canUseDOM} from 'shared/ExecutionEnvironment';
@@ -65,12 +66,6 @@ import {
   enqueueStateRestore,
   restoreStateIfNeeded,
 } from '../events/ReactDOMControlledComponent';
-import type {ReactPriorityLevel} from 'react-reconciler/src/ReactInternalTypes';
-import {
-  getCurrentUpdateLanePriority,
-  schedulerPriorityToLanePriority,
-  setCurrentUpdateLanePriority,
-} from 'react-reconciler/src/ReactFiberLane';
 
 setAttemptSynchronousHydration(attemptSynchronousHydration);
 setAttemptUserBlockingHydration(attemptUserBlockingHydration);
@@ -124,18 +119,6 @@ function createPortal(
 function scheduleHydration(target: Node) {
   if (target) {
     queueExplicitHydrationTarget(target);
-  }
-}
-
-// TODO: Remove this once callers migrate to alternatives.
-// This should only be used by React internals.
-function runWithPriority<T>(priority: ReactPriorityLevel, fn: () => T) {
-  const previousPriority = getCurrentUpdateLanePriority();
-  try {
-    setCurrentUpdateLanePriority(schedulerPriorityToLanePriority(priority));
-    return fn();
-  } finally {
-    setCurrentUpdateLanePriority(previousPriority);
   }
 }
 
@@ -215,7 +198,6 @@ export {
   createBlockingRoot,
   flushControlled as unstable_flushControlled,
   scheduleHydration as unstable_scheduleHydration,
-  runWithPriority as unstable_runWithPriority,
   // Disabled behind disableUnstableRenderSubtreeIntoContainer
   renderSubtreeIntoContainer as unstable_renderSubtreeIntoContainer,
   // Disabled behind disableUnstableCreatePortal
@@ -224,6 +206,9 @@ export {
   unstable_createPortal,
   // enableCreateEventHandleAPI
   createEventHandle as unstable_createEventHandle,
+  // TODO: Remove this once callers migrate to alternatives.
+  // This should only be used by React internals.
+  runWithPriority as unstable_runWithPriority,
 };
 
 const foundDevTools = injectIntoDevTools({
