+++ packages/react-dom/src/events/ReactDOMEventReplaying.js
@@ -18,7 +18,6 @@ import {
   enableDeprecatedFlareAPI,
   enableSelectiveHydration,
   enableModernEventSystem,
-  decoupleUpdatePriorityFromScheduler,
 } from 'shared/ReactFeatureFlags';
 import {
   unstable_runWithPriority as runWithPriority,
