+++ packages/react-reconciler/src/ReactFiberReconciler.js
@@ -50,6 +50,7 @@ import {
   focusWithin as focusWithin_old,
   observeVisibleRects as observeVisibleRects_old,
   registerMutableSourceForHydration as registerMutableSourceForHydration_old,
+  runWithPriority as runWithPriority_old,
 } from './ReactFiberReconciler.old';
 
 import {
@@ -88,6 +89,7 @@ import {
   focusWithin as focusWithin_new,
   observeVisibleRects as observeVisibleRects_new,
   registerMutableSourceForHydration as registerMutableSourceForHydration_new,
+  runWithPriority as runWithPriority_new
 } from './ReactFiberReconciler.new';
 
 export const createContainer = enableNewReconciler
@@ -190,7 +192,9 @@ export const focusWithin = enableNewReconciler
 export const observeVisibleRects = enableNewReconciler
   ? observeVisibleRects_new
   : observeVisibleRects_old;
-
 export const registerMutableSourceForHydration = enableNewReconciler
   ? registerMutableSourceForHydration_new
   : registerMutableSourceForHydration_old;
+export const runWithPriority = enableNewReconciler
+  ? runWithPriority_new
+  : runWithPriority_old;
